Improve scenario testing error messages (#7591)

+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/SettingsResources.scala
@@ -96,7 +96,7 @@ final case class EmptyDeploymentCommentSettingsError(message: String) extends Ex
 
 @JsonCodec final case class IntervalTimeSettings(processes: Int, healthCheck: Int)
 
-@JsonCodec final case class TestDataSettings(maxSamplesCount: Int, testDataMaxLength: Int, resultsMaxBytes: Int)
+@JsonCodec final case class TestDataSettings(maxSamplesCount: Int, testDataMaxLength: Int, resultsMaxBytes: Long)
 
 object TopTabType extends Enumeration {
